chores: dispatched moved to component call
- added seperate test for action and reducer
