+++ src/components/home/actions/appActions.js
@@ -1,68 +1,30 @@
-import { setNewWeb3, getWeb3, fetchNetwork } from "services/web3";
 import { appTypes } from "components/home/reducer/constants";
 
-export const updateNetworkId = async (dispatch, store) => {
-  try {
-    const provider = await getSelectedWeb3(dispatch, true, store);
-    const network = await fetchNetwork(provider);
-    const networkIdVerbose = matchNetwork(network.chainId);
-    return dispatch({
-      type: appTypes.UPDATE_NETWORK_ID_SUCCESS,
-      payload: {
-        network: network,
-        networkId: network.chainId,
-        networkIdVerbose
-      }
-    });
-  } catch (e) {
-    return dispatch({
-      type: appTypes.UPDATE_NETWORK_ID_FAILURE,
-      payload: e
-    });
-  }
-};
-
-export function matchNetwork(networkId) {
-  const networkIdVerbose = {
-    1: "Homestead",
-    2: "Morden",
-    3: "Ropsten",
-    4: "Rinkeby",
-    42: "Kovan"
+export function updateNetwork() {
+  return {
+    type: appTypes.UPDATE_NETWORK_ID
   };
-  return networkIdVerbose[networkId] || `Custom Network: ${networkId}`;
 }
 
-export const getSelectedWeb3 = async (dispatch, getNew = false, store) => {
-  const networkPending = await getNetworkPending(store);
-  if (networkPending && !getNew) {
-    // block if there's a network update pending
-    await dispatch({ type: appTypes.UPDATE_NETWORK_ID_SUCCESS });
-  }
-  const network = await getNetwork(store);
-  const web3 = (await getNew) ? setNewWeb3(network) : getWeb3(); // update web3 only if requested specifically
-  return web3;
-};
-
-export function foundNewBlock(dispatch, payload) {
-  return dispatch({
+export function foundNewBlock(payload) {
+  return {
     type: appTypes.NEW_BLOCK,
     payload
-  });
+  };
 }
 
-export function announceMinedTransaction(dispatch, payload) {
-  return dispatch({
+export function announceMinedTransaction(payload) {
+  return {
     type: appTypes.TRANSACTION_MINED,
     payload
-  });
+  };
 }
 
-export function removeTxFromPollingList(dispatch, payload) {
-  return dispatch({
+export function removeTxFromPollingList(payload) {
+  return {
     type: appTypes.TX_POLLING_REMOVE,
     payload
-  });
+  };
 }
 
 // Selectors
