+++ src/components/renderer/actions/index.js
@@ -1,6 +1,6 @@
 import { types } from "components/renderer/reducer/constants";
 
-const { DOCUMENT_TEMPLATE_REGISTER, DOCUMENT_TEMPLATE_SELECT_TAB } = types;
+const { DOCUMENT_TEMPLATE_REGISTER } = types;
 
 export function getDocument(store) {
   return store.document.docModified;
@@ -14,18 +14,11 @@ export function getTemplates(store) {
   return store.renderer.templates;
 }
 
-export function registerTemplates(dispatch, payload) {
-  return dispatch({
+export function registerTemplates(payload) {
+  return {
     type: DOCUMENT_TEMPLATE_REGISTER,
     payload
-  });
-}
-
-export function selectTemplateTab(dispatch, payload) {
-  return dispatch({
-    type: DOCUMENT_TEMPLATE_SELECT_TAB,
-    payload
-  });
+  };
 }
 
 export function getActiveTemplateTab(store) {
