+++ src/components/renderer/reducer/constants.js
@@ -1,12 +1,10 @@
 // Actions
 export const types = {
-  DOCUMENT_TEMPLATE_REGISTER: "DOCUMENT_TEMPLATE_REGISTER",
-  DOCUMENT_TEMPLATE_SELECT_TAB: "DOCUMENT_TEMPLATE_SELECT_TAB"
+  DOCUMENT_TEMPLATE_REGISTER: "DOCUMENT_TEMPLATE_REGISTER"
 };
 
 export const initialState = {
-  templates: null,
-  activeTemplateTab: 0
+  templates: null
 };
 
 export default types;
